initial QT gui

+++ remotecontrol2/tests/utils.py
@@ -36,46 +36,64 @@ class DebugPortConnection:
 class IntegrationTestBase(unittest.TestCase):
     maxDiff = None
     def setUp(self):
-        root_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-        self.source_dir = os.environ.get("CMAKE_SOURCE_DIR", os.path.join(root_path, "tests"))
-        self.binary_dir = os.path.join(
-            os.environ.get("CMAKE_BINARY_DIR", os.path.join(root_path, "build")),
-            "remotecontrol2")
+        self.server_process = None
+        self.client_process = None
+        self.client_process_thread = None
+        self.server_process_thread = None
 
-        server_executable_path = os.path.join(self.binary_dir, "server", "server")
-        if not os.path.exists(server_executable_path):
-            server_executable_path += ".exe"
-        self.assertTrue(os.path.exists(server_executable_path), f"The server_path '{server_executable_path}' was not found!")
-        self.server_process = subprocess.Popen(
-            [server_executable_path],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            bufsize=1, universal_newlines=True)
+        try:
+            root_path = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
+            self.source_dir = os.environ.get("CMAKE_SOURCE_DIR", os.path.join(root_path, "tests"))
+            self.binary_dir = os.path.join(
+                os.environ.get("CMAKE_BINARY_DIR", os.path.join(root_path, "build")),
+                "remotecontrol2")
 
-        self.server_process_thread = threading.Thread(
-            target=self.show_output,
-              args=(self.server_process, f"{self._testMethodName}() server output: "))
-        self.server_process_thread.start()
+            server_executable_path = os.path.join(self.binary_dir, "server", "server")
+            if not os.path.exists(server_executable_path):
+                server_executable_path += ".exe"
+            self.assertTrue(os.path.exists(server_executable_path), f"The server_path '{server_executable_path}' was not found!")
+            server_debug_port = 12345
+            self.server_process = subprocess.Popen(
+                [server_executable_path, "--debug-port", str(server_debug_port)],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+                bufsize=1, universal_newlines=True)
 
-        self.server_connection = DebugPortConnection("localhost", 12345)
+            self.server_process_thread = threading.Thread(
+                target=self.show_output,
+                args=(self.server_process, f"{self._testMethodName}() server output: "))
+            self.server_process_thread.start()
+
+            self.server_connection = DebugPortConnection("localhost", server_debug_port)
 
 
-        client_executable_path = os.path.join(self.binary_dir, "client", "client")
-        if not os.path.exists(client_executable_path):
-            client_executable_path += ".exe"
-        self.assertTrue(os.path.exists(client_executable_path), f"The client_path '{client_executable_path}' was not found!")
-        self.client_process = subprocess.Popen(
-            [client_executable_path],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            bufsize=1, universal_newlines=True)
+            client_executable_path = os.path.join(self.binary_dir, "client", "client")
+            if not os.path.exists(client_executable_path):
+                client_executable_path += ".exe"
+            self.assertTrue(os.path.exists(client_executable_path), f"The client_path '{client_executable_path}' was not found!")
+            client_debug_port = 21001
+            self.client_process = subprocess.Popen(
+                [client_executable_path,
+                 "--debug-port", str(client_debug_port),
+                 "--server-host", "localhost"
+                 ],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.STDOUT,
+                bufsize=1, universal_newlines=True)
 
-        self.client_process_thread = threading.Thread(
-            target=self.show_output,
-              args=(self.client_process, f"{self._testMethodName}() client output: "))
-        self.client_process_thread.start()
+            self.client_process_thread = threading.Thread(
+                target=self.show_output,
+                args=(self.client_process, f"{self._testMethodName}() client output: "))
+            self.client_process_thread.start()
 
-        self.client_connection = DebugPortConnection("localhost", 12346)
+            self.client_connection = DebugPortConnection("localhost", client_debug_port)
+        except Exception as e:
+            if self.client_process is not None:
+                self.client_process.kill()
+                self.client_process = None
+            if self.server_process is not None:
+                self.server_process.kill()
+                self.server_process = None
 
     def show_output(self, process, prefix):
         for line in process.stdout:
@@ -83,11 +101,15 @@ class IntegrationTestBase(unittest.TestCase):
             print(f"{prefix} - {line}")
 
     def tearDown(self):
-        self.client_process.kill()
-        self.server_process.kill()
+        if self.server_process is not None:
+            self.server_process.kill()
+            self.server_process.wait()
 
-        self.client_process.wait()
-        self.server_process.wait()
+        if self.client_process is not None:
+            self.client_process.kill()
+            self.client_process.wait()
 
-        self.client_process_thread.join()
-        self.server_process_thread.join()
+        if self.client_process_thread is not None:
+            self.client_process_thread.join()
+        if self.server_process_thread is not None:
+            self.server_process_thread.join()
